requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
distlib==0.4.0
Django==5.2.7
django-stubs==5.2.7
django-stubs-ext==5.2.7
djangorestframework==3.16.1
filelock==3.20.0
gunicorn==23.0.0
identify==2.6.15
idna==3.11
iniconfig==2.3.0
kombu==5.5.4
nodeenv==1.9.1
packaging==25.0
platformdirs==4.5.0
pluggy==1.6.0
pre_commit==4.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
redis==5.2.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
virtualenv==20.35.4
wcwidth==0.2.14
